mf_some/api/posts/views.py
from posts.models import Like, Post
from rest_framework.viewsets import ModelViewSet

from .serializers import LikeSerializer, PostSerializer


class PostViewSet(ModelViewSet):
    queryset = Post.objects.all()
    serializer_class = PostSerializer


class LikeViewSet(ModelViewSet):
    queryset = Like.objects.all()
    serializer_class = LikeSerializer
