requirements/base.txt
psycopg2-binary==2.7.5
Django==2.1.2
djangorestframework==3.8.2
django-improved-user==1.0.0
channels==2.1.3
channels-redis==2.3.0
bpython
